requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
Django==3.2.5
djangorestframework==3.12.4
idna==2.10
psycopg2==2.9.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.6
